Time: 388 ms (49.35%) | Memory: 211.6 MB (38.20%) - LeetSync

+++ 123-best-time-to-buy-and-sell-stock-iii/best-time-to-buy-and-sell-stock-iii.cpp
@@ -15,19 +15,19 @@ class Solution {
 public:
     int maxProfit(vector<int>& prices) {
         int n = prices.size();
-        vector<vector<vector<long>>> dp(n+1,vector<vector<long>>(2,vector<long>(3,0)));
+        vector<vector<vector<int>>> dp(n+1,vector<vector<int>>(2,vector<int>(3,0)));
 
         for(int i=n-1;i>=0;i--){
             for(int j=0;j<2;j++){
                 for(int k=1;k>=0;k--){
                     if(j == 0){
-                        long include = prices[i] + dp[i+1][1][k+1];
-                        long skip = dp[i+1][0][k];
+                        int include = prices[i] + dp[i+1][1][k+1];
+                        int skip = dp[i+1][0][k];
                         dp[i][j][k] = max(include,skip);
                     }
                     else{
-                        long include = -prices[i] + dp[i+1][0][k];
-                        long skip = dp[i+1][1][k];
+                        int include = -prices[i] + dp[i+1][0][k];
+                        int skip = dp[i+1][1][k];
                         dp[i][j][k] = max(include,skip);
                     }
                 }
